demo7-replay/service-collision-randomgenerator/requirements.txt
kafka-python==2.0.2
python-dotenv==0.19.0
pandas
numpy
